Revert "Add fabriceNode GET API support"
